fold VideoDriverCommon into VideoDriver
