+++ src/config/front-ends/headless/headless.h
@@ -6,10 +6,11 @@ class HeadlessVideoDriver : public Executor::VideoDriver
 {
     using VideoDriver::VideoDriver;
 
-    virtual void setColors(int num_colors, const Executor::vdriver_color_t *color_array) override;
     virtual bool setMode(int width, int height, int bpp,
                                 bool grayscale_p) override;
 
     virtual void runEventLoop() override {}
     virtual void endEventLoop() override {}
+protected:
+    virtual void requestUpdate() override {}
 };
